fix: divide test community user
